requirements.txt
aws-cdk-lib==2.20.0
constructs>=10.0.0,<11.0.0
boto3>=1.9.0,<2.0.0
